% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate.R
\name{animate_error_bar}
\alias{animate_error_bar}
\title{Create animated gif of error bars}
\usage{
animate_error_bar(
  mc_filtered,
  demographic = FALSE,
  moving_scale = FALSE,
  path = "./"
)
}
\arguments{
\item{demographic}{demographic to plot}

\item{mc}{An mc object from the mcdata package.}
}
\description{
Create animated gif of error bars
}
