% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_updates.R
\name{monthly_demographic_comparison}
\alias{monthly_demographic_comparison}
\title{Monthly demographic Table}
\usage{
monthly_demographic_comparison(
  mc_tidy,
  current_month = lubridate::floor_date(Sys.Date(), "month"),
  last_month = lubridate::floor_date(Sys.Date(), "month") - months(1)
)
}
\arguments{
\item{mc_tidy}{mcdata object}

\item{current_month}{(optional) Default = Current month. The later month used for comparison}

\item{last_month}{(optional) Default = Last month. The earlier month used for comparison}
}
\description{
Monthly demographic Table
}
