% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_updates.R
\name{nih_plots}
\alias{nih_plots}
\title{NIH Plots}
\usage{
nih_plots(prod = FALSE, upload = FALSE)
}
\arguments{
\item{prod}{TRUE or FALSE if being used in production environment}

\item{upload}{FALSE or GDrive ID for upload location}
}
\description{
NIH Plots
}
