% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{plot_age_barplot}
\alias{plot_age_barplot}
\title{Plot Age Barplot}
\usage{
plot_age_barplot(
  mc,
  title = "",
  percentage = TRUE,
  age_decade = FALSE,
  last_month = FALSE,
  path = FALSE
)
}
\arguments{
\item{mc}{An mc object from the mcdata package.}

\item{demographic}{A string of the demographic you wish to plot.}
}
\description{
Plot Age Barplot
}
