% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{plot_demographic_barplot}
\alias{plot_demographic_barplot}
\title{Generate a demographic barplot from an mc object}
\usage{
plot_demographic_barplot(mc, demographic, title = "", subset = FALSE)
}
\arguments{
\item{mc}{An mc object from the mcdata package.}

\item{demographic}{A string of the demographic you wish to plot.}
}
\description{
Generate a demographic barplot from an mc object
}
