% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{plot_geo}
\alias{plot_geo}
\title{Plot geographic participation data}
\usage{
plot_geo(mc, scale = "world")
}
\arguments{
\item{mc}{An mc object from the mcdata package.}

\item{scale}{"world" for world plot or "us" for US plot}
}
\description{
Plot geographic participation data
}
