% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{plot_model_results}
\alias{plot_model_results}
\title{Plot model results line plots by demographic}
\usage{
plot_model_results(
  mc,
  plain_title = "",
  demographic = "age",
  demographic_subset = FALSE,
  smooth = FALSE,
  game = "totalcorrect",
  model_results = NA
)
}
\arguments{
\item{mc}{An mc object from the mcdata package.}

\item{plain_title}{(optional) Title for plot. Default = ""}

\item{demographic}{(optional) Demographic to plot. Default = "age"}

\item{demographic_subset}{(optional) A vector containing the demographic values you wish to plot}

\item{smooth}{(optional) Default = FALSE, returning linear lines. If TRUE, smooth lines are returned}

\item{game}{(optional) Default = "totalcorrect". For SVRT results, use "medianSVRT_no_outliers"}

\item{model_results}{(optional) mc_model output object}
}
\description{
Plot model results line plots by demographic
}
