% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_updates.R
\name{plot_participation_barplot}
\alias{plot_participation_barplot}
\title{Participation N Barplot}
\usage{
plot_participation_barplot(mc, path = FALSE)
}
\arguments{
\item{mc}{mcdata object}
}
\description{
Participation N Barplot
}
