% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_updates.R
\name{plot_participation_line}
\alias{plot_participation_line}
\title{Participation N Line}
\usage{
plot_participation_line(mc)
}
\arguments{
\item{mc}{mcdata object}
}
\description{
Participation N Line
}
