% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{plot_percent_contactable}
\alias{plot_percent_contactable}
\title{Contactable Lineplot}
\usage{
plot_percent_contactable(mc, demographic = FALSE, demographic_subset = FALSE)
}
\arguments{
\item{mc}{An mc object from the mcdata package.}

\item{demographic}{(optional) Default = FALSE. A string of the demographic you wish to plot.}
}
\description{
Generate a line plot of the percentage of contactable participants from ages 18-90, optionally by a specified demographic
}
