% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_updates.R
\name{plot_projected_participants}
\alias{plot_projected_participants}
\title{Recruitment Projection Line}
\usage{
plot_projected_participants(
  mc,
  projection_start_date = "2020-01-01",
  end_date = "2030-01-01",
  goal = FALSE
)
}
\arguments{
\item{mc}{mcdata object}

\item{projection_start_date}{(optional) Default = "2020-01-01", Start date of projection}

\item{end_date}{(optional) Default = "2030-01-01", End date of projection}
}
\description{
Recruitment Projection Line
}
