% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplotter.R
\name{run_autoplotter}
\alias{run_autoplotter}
\title{Run the autoplotter}
\usage{
run_autoplotter(prod = FALSE, geo = TRUE, upload = FALSE)
}
\arguments{
\item{prod}{(optional) Default = FALSE. Indicates if this is running on production}

\item{geo}{(optional) Default = TRUE. Set to FALSE to remove sf dependency}

\item{upload}{(optional) Default = FALSE. To upload output to drive, set this equal to drive folder ID}
}
\description{
Run the autoplotter
}
