% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_results.R
\name{split_demo}
\alias{split_demo}
\title{Split model results labels for plotting}
\usage{
split_demo(model)
}
\arguments{
\item{model}{mcmodel results}
}
\description{
Split model results labels for plotting
}
